chore: encapsulate mqtt part in its own iot module, which will save to DB in the future
